requirements.txt
fastapi
uvicorn[standard]
psycopg2
sqlalchemy
python-jose
passlib
python-multipart